customer_obfuscation/requirements.txt
absl-py==0.10.0
astor==0.8.1
boto3==1.9.117
certifi==2020.6.20
cycler==0.10.0
gast==0.3.3
google-cloud-storage==1.16.1
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
importlib-metadata==1.7.0
ipdb==0.12
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==2.2.5
numpy==1.16.6
opencv-python==4.1.0.25
opt-einsum==2.3.2
Pillow==6.0.0
protobuf==3.13.0
pyparsing==2.4.7
python-dateutil==2.8.1
PyYAML==5.1
redis==4.2.0
requests==2.21.0
six==1.15.0
tensorboard==2.2.0
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
Werkzeug==1.0.1
wrapt==1.12.1
zipp==1.2.0